+++ backend/plugin/vcs/bitbucket/bitbucket_test.go
@@ -1,4 +1,4 @@
-package bitbucketcloud
+package bitbucket
 
 import (
 	"context"
